src/requirements.txt
# PDF Parser 模块依赖
PyPDF2==3.0.1
pdfplumber==0.10.3
pandas>=2.3.0
openpyxl>=3.1.5
python-docx>=1.2.0
Pillow>=11.3.0
pypdfium2>=4.30.0

# OCR 模块依赖
openai>=1.0.0
requests>=2.32.0
